Fix REST API on backend side

+++ backend/issue_service/src/main/java/com/example/issue_service/IssueController.java
@@ -32,16 +32,20 @@ public class IssueController {
         this.issueRepository.save(issue);
     }
 
-    @PutMapping
-    @ResponseStatus(HttpStatus.ACCEPTED)
+    @PutMapping(path = "/{id}")
+    @ResponseStatus(HttpStatus.OK)
     public void update(@RequestBody Issue issue) {
         issue.setUpdatedAt(LocalDateTime.now());
         this.issueRepository.save(issue);
     }
 
-    @DeleteMapping
-    @ResponseStatus(HttpStatus.ACCEPTED)
-    public void delete(@PathVariable long id) {
-        this.issueRepository.deleteById(id);
+    @DeleteMapping(path = "/{id}")
+    public ResponseEntity<String> delete(@PathVariable long id) {
+        Optional<Issue> issue = this.issueRepository.findById(id);
+        if(issue.isPresent()) {
+            this.issueRepository.deleteById(id);
+            return new ResponseEntity<String>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<String>(String.format("issue with id=%s not found!", id), HttpStatus.NOT_FOUND);
     }
 }
